initial mouse support
This patch adds support for mouse input and cursor hiding. You can
scroll by holding the right button pressed and releasing it after moving
it the distance you want to scroll. Zooming can be done by using the
mousewheel. The cursor will be hidden after not moving it for 500 ms.

Signed-off-by: Florian Tobias Schandinat <FlorianSchandinat@gmx.de>