make startup width & height configurable
Currently SDL creates the window at maximum width and height which is
nearly always a bit too big as it does not take menu bars and similiar
into account. Unfortunately SDL provides no easy way to allow the window
manager to select it. Provide the command line arguments -width= and
-height= to get around this. This is only a little improvement as long
as session or configuration files are not implemented.
On the way the command line option -fullscreen was added to allow
fullscreen operation.

Signed-off-by: Florian Tobias Schandinat <FlorianSchandinat@gmx.de>