initial configuration file and scheme support
This adds basic support for configuration files and schemes. Currently
only width, height and fullscreen settings are supported. Although it
should work okay don't expect it to be stable yet. The basic
configuration infrastructure is planned to become stable with the next
release (version 0.2). Full and stable configuration file options are
expected between version 0.5 and 1.0.

Signed-off-by: Florian Tobias Schandinat <FlorianSchandinat@gmx.de>