implement offline help
This patch implements a usefull "green -help". It only lists the
preferred option style although a '--' option prefix is also allowed as
well as for "help" and "version" the short forms "h" and "v" exist.
Options that are not listed are unstable and might go away in the near
future.

Signed-off-by: Florian Tobias Schandinat <FlorianSchandinat@gmx.de>